% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_diameter_zone.R
\name{add_diameter_zone}
\alias{add_diameter_zone}
\title{add_diameter_zone}
\usage{
add_diameter_zone(
  diam_min,
  diam_max,
  coeff_traj = c(a = 13.532, b = -1.461, a0 = 14.21, b0 = -1.79),
  ...
)
}
\arguments{
\item{diam_min}{numeric corresponding of the minimal harvestable diameter}

\item{diam_max}{numeric corresponding of the minimal harvestable diameter}

\item{coeff_traj}{coefficient a, b, a0, and b0 of the mortality trajectory (cf : Ningre et al "Trajectoires d auto-eclaircie du Douglas en France)}

\item{...}{other argument passed on to layer (ex : color, linetype, ...)}
}
\value{
ggplot object highlighting zone of harvestable diameter
}
\description{
add_diameter_zone
}
\examples{
\dontrun{add_harvestable_diameter(diam_min = 40, diam_max = 45, fill = "orange", alpha = 0.5)}

}
