% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_isolines.R
\name{add_isolines}
\alias{add_isolines}
\title{add_isolines}
\usage{
add_isolines(
  stand_data,
  variable = Hdom,
  iso_values = c(26, 33, 40),
  label_position = "top",
  nudge_label = c(x = 0, y = 0),
  show_point = FALSE,
  ...
)
}
\arguments{
\item{stand_data}{data.frame from the use of "import_stand_data" function}

\item{variable}{one variable from the stand_data (ex : Hdom, Gha, ...)}

\item{iso_values}{a vector of isoline values (ex : c(10,20,30,40,50))}

\item{label_position}{choose the position of the label with "top", "bottom", "none"}

\item{nudge_label}{more accurate positionning of labels by offsetting them in x and y direction (ex : -0.1,0.2)}

\item{show_point}{if TRUE point used to find isolines are shown}

\item{...}{other argument passed on to layer (ex : color, linetype, ...)}
}
\value{
ggplot object with isolines
}
\description{
add_isolines
}
\examples{
\dontrun{add_isolines()}

}
