% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRDI.R
\name{add_RDI_line}
\alias{add_RDI_line}
\title{add_RDI_line}
\usage{
add_RDI_line(
  RDI = 0.15,
  coeff_traj = c(a = 13.532, b = -1.461, a0 = 14.21, b0 = -1.79),
  ...
)
}
\arguments{
\item{RDI}{numeric between 0 and 1}

\item{coeff_traj}{coefficient a, b, a0, and b0 of the mortality trajectory (cf : Ningre et al "Trajectoires d auto-eclaircie du Douglas en France)}

\item{...}{other argument passed on to layer (ex : color, linetype, ...)}
}
\value{
ggplot objetc correpsonding to an RDI line
}
\description{
add_RDI_line
}
\examples{
\dontrun{add_RDI_line(RDI = 0.15)}

}
