% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_growth.R
\name{calculate_growth}
\alias{calculate_growth}
\title{calculate_growth}
\usage{
calculate_growth(stand_data, variable, growth_type)
}
\arguments{
\item{stand_data}{data.frame from the use of "import_stand_data" function}

\item{variable}{variable on which the growth will be calculated}

\item{growth_type}{should be "annual" or "average" which mean from the start of the stand}
}
\value{
add a column to data.frame with the calculated growth
}
\description{
calculate_growth
}
