% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_traj.R
\name{calculate_traj}
\alias{calculate_traj}
\title{calculate_traj}
\usage{
calculate_traj(
  cg_range,
  density_init = 1000,
  coeff_traj = c(a = 13.532, b = -1.461, a0 = 14.21, b0 = -1.79)
)
}
\arguments{
\item{cg_range}{Range of calculation for the trajectory}

\item{density_init}{A vector of initial densities (numeric)}

\item{coeff_traj}{Coefficient a, b, a0, and b0 of the mortality trajectory (cf : Ningre et al "Trajectoires d auto-eclaircie du Douglas en France)}
}
\value{
Data.frame with trajectory calculated for the cg_range
}
\description{
calculate_traj
}
\examples{
\dontrun{calculate_traj()}

}
