% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateRV.R
\name{GenerateRV}
\alias{GenerateRV}
\title{GenerateRV generic}
\usage{
GenerateRV(obj, n, ...)
}
\arguments{
\item{obj}{object from which to sample.}

\item{n}{number of items to sample.}

\item{\dots}{further arguments for methods.}
}
\description{
A generic to sample random variates from an object.
}
\author{
Berwin A. Turlach \email{berwin.turlach@gmail.com}
}
\keyword{distribution}
