% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stand_data.R
\name{import_stand_data}
\alias{import_stand_data}
\title{import_stand_data}
\usage{
import_stand_data(filepath)
}
\arguments{
\item{filepath}{filepath of the output folder from SimCop. You need at least "stand_data.csv" and "parameter_scenarii.csv"}
}
\value{
data.frame of stand informations
}
\description{
import_stand_data
}
