% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thinning_coefficient.R
\name{plot_thinning_coefficient}
\alias{plot_thinning_coefficient}
\title{plot_thinning_coefficient}
\usage{
plot_thinning_coefficient(stand_data, remove_theme = FALSE)
}
\arguments{
\item{stand_data}{data.frame from the use of "import_stand_data" function AND add_ecl_data set to TRUE}

\item{remove_theme}{If TRUE, already created theme for the graph is remove}
}
\value{
ggplot object representing variation of K parameters
}
\description{
plot_thinning_coefficient
}
