% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traj.R
\name{plotTraj}
\alias{plotTraj}
\title{plot_traj}
\usage{
plotTraj(
  density_init = 1000,
  coeff_traj = c(a = 13.532, b = -1.461, a0 = 14.21, b0 = -1.79),
  scale = c(xmin = 15, xmax = 500, ymin = 50, ymax = 5000),
  ...,
  base_theme = TRUE
)
}
\arguments{
\item{density_init}{A vector of initial densities (numeric)}

\item{coeff_traj}{Coefficient a, b, a0, and b0 of the trajectory (cf : Ningre et al "Trajectoires d’autoéclaircie du Douglas en France)}

\item{scale}{A vector compose of xmin, xmax, ymin and ymax}

\item{...}{Param to give to th ggplot (color, linetype, alpha, ...)}

\item{base_theme}{If TRUE, allows to apply a theme already created for the graph}
}
\value{
A ggplot object
}
\description{
plot_traj
}
\examples{
\dontrun{plot_traj()}

}
