% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ADCube}
\alias{print.ADCube}
\title{Print method for cubes}
\usage{
\method{print}{ADCube}(x, ...)
}
\arguments{
\item{x}{an object of \code{\link{class}} \sQuote{ADCube}.}

\item{\dots}{arguments passed on to \code{base::print}.}
}
\description{
\code{\link{print}} method for \sQuote{ADCube} objects.
}
\details{
Currently, the component \sQuote{\code{val}} of the \sQuote{ADCube} object is printed using \code{base::print}.
}
\keyword{internal}
