% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SimCop}
\alias{print.SimCop}
\title{Print basic information on a copula}
\usage{
\method{print}{SimCop}(x, ...)
}
\arguments{
\item{x}{an object of \code{\link{class}} \sQuote{SimCop}.}

\item{\dots}{not used.}
}
\description{
Prints basic information on a copula created with the methods in this package.
}
\author{
Berwin A. Turlach <berwin.turlach@gmail.com>
}
\keyword{print}
