% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_non_simulated_zone.R
\name{add_non_simulated_zone}
\alias{add_non_simulated_zone}
\title{add_non_simulated_zone}
\usage{
add_non_simulated_zone(
  stand_data,
  coeff_traj = c(a = 13.532, b = -1.461, a0 = 14.21, b0 = -1.79)
)
}
\arguments{
\item{stand_data}{data.frame from the use of "import_stand_data" function}

\item{coeff_traj}{coefficient a, b, a0, and b0 of the mortality trajectory (cf : Ningre et al "Trajectoires d auto-eclaircie du Douglas en France)}
}
\value{
A ggplot object corresponding to the zone of non-simulated data
}
\description{
add_non_simulated_zone
}
