% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stand_data.R
\name{import_stand_data}
\alias{import_stand_data}
\title{import_stand_data}
\usage{
import_stand_data(filepath, after_mortality_only = TRUE)
}
\arguments{
\item{filepath}{filepath of the output folder from SimCop. You need at least "stand_data.csv" and "parameter_scenarii.csv"}

\item{after_mortality_only}{if TRUE, the result only keep values after mortality}
}
\value{
data.frame of stand informations
}
\description{
import_stand_data
}
